requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
gotrue==2.4.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
postgrest==0.16.2
psycopg2==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.1
supafunc==0.4.5
typing_extensions==4.10.0
websockets==12.0
Werkzeug==3.0.1
